pyproject.toml
[project]
name = "book-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "sqlalchemy",
    "alembic",
    "sqlmodel",
    "pydantic",
    "pydantic-settings",
    "asyncpg",
    "python-dotenv",
    "passlib>=1.7.4",
    "bcrypt>=5.0.0",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt[crypto]>=2.10.1",
    "redis>=7.1.0",
]
